+++ src/Constraint/Map.php
@@ -4,27 +4,41 @@
 namespace Stefmachine\Validation\Constraint;
 
 
+use InvalidArgumentException;
+use Stefmachine\Validation\Constraint\Traits\ErrorMessageTrait;
 use Stefmachine\Validation\ConstraintInterface;
-use Stefmachine\Validation\Errors;
+use Traversable;
+use UnexpectedValueException;
 
 class Map implements ConstraintInterface
 {
-    const ERROR_ARRAY = 'not_array';
-    const ERROR_EXTRA = 'extra_field';
-    
     /** @var ConstraintInterface[] */
     protected $constraints;
     /** @var bool */
     protected $allowExtra;
     
-    public function __construct(array $_map = array(), bool $_allowExtra = false)
+    use ErrorMessageTrait;
+    
+    public function __construct(array $_map, ?string $_errorMessage = null)
     {
-        $this->allowExtra = $_allowExtra;
+        $this->allowExtra = false;
         $this->constraints = array();
         
+        $this->setErrorMessage($_errorMessage);
+        
         foreach ($_map as $key => $validation){
+            if(!$validation instanceof ConstraintInterface){
+                throw new InvalidArgumentException("Expected array of ".ConstraintInterface::class.".");
+            }
+            
             $this->set($key, $validation);
         }
+    }
+    
+    private function set(string $_key, ConstraintInterface $_constraint): Map
+    {
+        $this->constraints[$_key] = $_constraint;
+        return $this;
     }
     
     public function allowExtra(): Map
@@ -33,42 +47,50 @@ class Map implements ConstraintInterface
         return $this;
     }
     
-    public function set(string $_key, ConstraintInterface $_constraint): Map
+    public function disallowExtra(): Map
     {
-        $this->constraints[$_key] = $_constraint;
+        $this->allowExtra = false;
         return $this;
     }
     
-    public function validate($_value): Errors
+    public function validate($_value)
     {
-        if(!is_array($_value) && !$_value instanceof \Traversable){
-            return Errors::from(self::ERROR_ARRAY);
+        if(!is_array($_value) && !$_value instanceof Traversable){
+            throw new UnexpectedValueException("Value is not traversable.");
         }
         
-        $errors = Errors::none();
+        $errors = [];
         $missingKeys = array_keys($this->constraints);
         foreach ($_value as $key => $value){
             $index = array_search($key, $missingKeys);
             if($index !== false){
                 unset($missingKeys[$index]);
     
-                $errors->merge(
-                    $this->constraints[$key]->validate($value),
-                    "{$key}."
-                );
+                $errors[] = $this->constraints[$key]->validate($value);
             }
             elseif (!$this->allowExtra){
-                return $errors->add(self::ERROR_EXTRA);
+                return $this->getError();
             }
         }
         
         foreach ($missingKeys as $key){
-            $errors->merge( // Giving null to trigger validations
-                $this->constraints[$key]->validate(null),
-                "{$key}."
-            );
+            $errors[] = $this->constraints[$key]->validate(null);
         }
         
-        return $errors;
+        $hasError = false;
+        $errorMessages = array();
+        foreach ($errors as $thisError){
+            if($thisError !== true){
+                $hasError = true;
+                
+                if(is_string($thisError)){
+                    $errorMessages[] = $thisError;
+                }
+            }
+        }
+        
+        $errorMessages = empty($errorMessages) ? false : implode(' ', $errorMessages);
+        
+        return !$hasError ?: $this->getError() ?: $errorMessages;
     }
 }