+++ src/Constraint/MinCount.php
@@ -4,32 +4,34 @@
 namespace Stefmachine\Validation\Constraint;
 
 
+use Countable;
+use InvalidArgumentException;
+use Stefmachine\Validation\Constraint\Traits\ErrorMessageTrait;
 use Stefmachine\Validation\ConstraintInterface;
-use Stefmachine\Validation\Errors;
-use Stefmachine\Validation\Helper\ErrorMaker;
+use UnexpectedValueException;
 
 class MinCount implements ConstraintInterface
 {
-    const ERROR_NOT_COUNTABLE = 'not_countable';
-    const ERROR_MIN_COUNT = 'min_count';
-    
     protected $minCount;
     
-    public function __construct(int $_minCount)
+    use ErrorMessageTrait;
+    
+    public function __construct(int $_minCount, ?string $_errorMessage = null)
     {
+        if($_minCount < 0){
+            throw new InvalidArgumentException("Min count cannot be less than 0.");
+        }
+        
         $this->minCount = $_minCount;
+        $this->setErrorMessage($_errorMessage);
     }
     
-    public function validate($_value): Errors
+    public function validate($_value)
     {
-        if(!is_array($_value) || $_value instanceof \Countable){
-            return Errors::from(self::ERROR_NOT_COUNTABLE);
+        if(!is_array($_value) && !$_value instanceof Countable){
+            throw new UnexpectedValueException("Value is not an array or countable.");
         }
-        
-        if(count($_value) < $this->minCount){
-            return Errors::from(ErrorMaker::makeError(self::ERROR_MIN_COUNT, ['min' => $this->minCount]));
-        }
-        
-        return Errors::none();
+    
+        return count($_value) >= $this->minCount ?: $this->getError();
     }
 }