+++ CHANGELOG.md
@@ -2,6 +2,22 @@
 
 All notable changes to this project will be documented in this file.
 
+## [0.4.0]
+### Changed
+- The whole API now works by returning true/false or message instead of an Errors object
+
+### Added
+- Unit tests
+
+### Removed
+- Translations
+- Error Parser
+- Errors class
+
+## [0.3.0]
+### Added
+- Translations
+
 ## [0.2.2]
 ### Added
 - Documentation